requirements.txt
django==4.1
pytz==2022.2.1
uritemplate==4.1.1
PyYAML==6.0
Markdown==3.4.1
Pygments==2.13.0
django-filter==22.1
django-guardian==v2.4.0
djangorestframework==3.12.4
mysqlclient==2.1.1
django-localflavor==3.1
drf-spectacular==0.24.2
django-cors-headers==3.13.0